Fix removal of padding on decryption.
Trouble is that we only get passed the EOF flag when we get passed a
zero length buffer, which means it's essentially to late, as we have
written out the padding already. So, try to detect when the last real
block is passed ...

Also make treatment of keys and IVs more consistent:
* Keys are considered 32bit integers, thus outputting as numbers rather than
  byte streams.
* IVs are considered bytestream data.

This does not yet make us openssl compatible and we may reverse keys now
incorrectly byteswap twice on little endian machines -- needs more
investigation.