Optimize ARM8 AES with 4x pipelined processing.
We can start aese insns without waiting for previous ones to complete.
We do 4 at a time before we do the subsequent aesmc that have a data
dependency, thus hiding some of the latency.

This achieves < 2 cycles for bulk ECB encryption on ARM A5x, ~30% speedup.

Same trick speeds up ECB decryption and CBC decryption (but not CBC encryption).
We also use it for CTR speedup (~15%), but this can be further improved.