Fix err pos reporting for reverse; fix EOF faults.
When finding fault (via injection or real) in the block before EOF,
we would produce a too large output file. Handle this case.

When reverse copying, the faults would be reported to be located in the
block after the one really affected. Fix reporting.

TODO: Fault inection fails on the first block (last block in file) when
reverse copying.