Major cleanup for write retry handling.
The old code did not properly account for plugins that can change
the size of a block. Also the retry logic was implemented at a too
high layer, so we ended up rewriting the secondary files again,
calling the plugins again etc.
The new logic does account for this and is a lot cleaner in general;
we now also advance positions correctly in case of write errors even
if the plugins change the size ...
Actually, if it weren't for documentation, the changes would have
reduced the source code size :-)