An awfully hacky solution that currently works.
We derive the offsets of struct fields by calculating them,
using our knowledge of the struct definition and the knowledge
of C rules how structs are laid out.
This will break if fields get inserted before final_used.
Not silently though, there are two asserts that make this hack
at least half-way safe.

Alternative would be to store a copy of the context and restore
it when needed.