Pass correct MACH=armv6 and armv7 to main Makefile.
When building binaries with the Android NDK, we still use the main
Makefile that selects which machine-specific objects get built to
speed up detecting zeroed blocks and crypto. There was a mismatch
where the Android Makefiles would pass MACH=arm where the main
Makefile differentiates between armv6 and armv7.

The old API9 Android binaries are built for armv6 (so we have max
compatibility), whereas the API16 PIE binaries require armv7.
(64bit is API20 and did not need to be changed.)

Signed-off-by: Kurt Garloff <kurt@garloff.de>