Use better insn scheduling in aesni CTR.
Hard to measure, it looks a tiny bit better, but it's well within
the noise (measured on Skylake).

Also fixed inconsistency where the CTR value would not have been
incremented for the last block if it was incomplete. This was
different from the other implementations (C, ARMv8).
In the context of dd_rescue it was cosmetic however.

Signed-off-by: Kurt Garloff <kurt@garloff.de>