Let's not discount the last incomplete CTR block.
We can't easily do this in openSSL.
Being consistent with openSSL even in corner cases is good,
so let's increase our consistency. Namely, when the last blocks
need padding, let's ensure our IV state is the same as openSSL's.
Previously we have not done the final increase of ctr for the
last incomplete block. This with the idea that attaching another
partial block seemlessly would require using the same IV again ...
Good thought, but consistency with openSSL seems more important.

Signed-off-by: Kurt Garloff <kurt@garloff.de>