Compile aesni twice, once with -mavx2, once without.
This is unfortunately the only way to avoid SIGILL, as we
need to compile with -mavx(2) to use the AVX intrinsics,
but then the compiler prefers VEX instructions ...

So we're compiling the file twice now and making everything
static (except the method list, which we dynamically name).

Selection will be done at runtime, based on have_avx2 detection.
For libddr_crypt, we can use engine=vaesni/saesni to override
the automated selection.

Signed-off-by: Kurt Garloff <kurt@garloff.de>